Move MultiDirsSelectDialog class to picard.ui.util
