+++ picard/ui/util.py
@@ -93,3 +93,22 @@ class ButtonLineEdit(QtGui.QLineEdit):
         self.clear_button.setVisible(text != "")
         left, top, right, bottom = self._margins
         self.setTextMargins(left, top, right + self.clear_button.width(), bottom)
+
+
+class MultiDirsSelectDialog(QtGui.QFileDialog):
+
+    """Custom file selection dialog which allow the selection
+    of multiple directories.
+    Depending on the platform, dialog may fallback on non-native.
+    """
+
+    def __init__(self, *args):
+        super(MultiDirsSelectDialog, self).__init__(*args)
+        self.setFileMode(self.Directory)
+        self.setOption(self.ShowDirsOnly)
+        if sys.platform == "darwin":
+            # The native dialog doesn't allow selecting >1 directory
+            self.setOption(self.DontUseNativeDialog)
+        for view in self.findChildren((QtGui.QListView, QtGui.QTreeView)):
+            if isinstance(view.model(), QtGui.QFileSystemModel):
+                view.setSelectionMode(QtGui.QAbstractItemView.ExtendedSelection)
